refactor: move utxo coins to utxo/coins
Ticket: BTC-1092

TICKET: BTC-1092
