+++ modules/statics/src/utxo/index.ts
@@ -1,1 +1,2 @@
 export * from './utxo';
+export * from './coins';
